Unify the addDate for the backend and frontend spoonform
